apply feed
