+++ src/lib/types.ts
@@ -36,7 +36,7 @@ import { SdkActionTypes } from './dux/sdk/actionTypes';
 import { ReconnectType } from './hooks/useConnect/types';
 import { SBUGlobalPubSub } from './pubSub/topics';
 import { EmojiManager } from './emojiManager';
-import {MessageTemplatesInfo, ProcessedMessageTemplate, WaitingTemplateKeyData} from './dux/appInfo/initialState';
+import { MessageTemplatesInfo, ProcessedMessageTemplate, WaitingTemplateKeyData } from './dux/appInfo/initialState';
 import { AppInfoActionTypes } from './dux/appInfo/actionTypes';
 
 // note to SDK team:
